"Feat: Completed all projects error cases"

+++ pkg/projects/delete_project.go
@@ -9,24 +9,40 @@ import (
 
 // @Summary Delete a specific Project
 // @Description DELETE a specific project. For the request to be met, the "id_projeto" are required
-// @Param        ID   						path      	int  	true  	"Projeto ID"
+// @Param        id   		path      	int  	true  	"Projeto ID"
 // @Accept json
 // @Produce json
 // @Success 200 {array} models.Projeto
-// @Failure 400,404 {string} string "error"
+// @Failure 400 {array} models.Error400Delete
+// @Failure 404 {array} models.Error404Delete
 // @Tags Projects
 // @Router /projetos/{id} [delete]
 func (h handler) DeleteProject(c *gin.Context) {
 	id := c.Param("id")
-
+	var IdExist int
 	var projeto models.Projeto
 
-	if result := h.DB.First(&projeto, id); result.Error != nil {
+	// Search the database if the person with the selected id exists
+	if result := h.DB.Raw("select count(*) from projetos where id_projeto = ?", id).Scan(&IdExist); result.Error != nil {
 		c.AbortWithError(http.StatusNotFound, result.Error)
 		return
 	}
 
-	h.DB.Delete(&projeto)
+	// if it exists then delete it
+	if IdExist == 1 {
+		if result := h.DB.First(&projeto, id); result.Error != nil {
+			c.AbortWithError(http.StatusNotFound, result.Error)
+			return
+		}
+	
+		h.DB.Delete(&projeto)
+	
+		c.Status(http.StatusOK)
+	} else {
+		c.JSON(400, gin.H{
+			"message": "Unable to delete. non-existent ID " ,
+		})
+		return
+	}
 
-	c.Status(http.StatusOK)
 }
